+++ NEWS.md
@@ -1,3 +1,4 @@
-# semhelpinghands 0.0.0.9001
+# semhelpinghands 0.0.1.9001 (WIP)
 
-* Initial commits.
+- Add `standardizedSolution_boot_ci()` for forming bootstrap percentile
+  confidence intervals for standardized solution in an `lavaan` output.